+++ ExtractTable/config.py
@@ -8,6 +8,7 @@ class HOST:
     VALIDATOR = 'validator.extracttable.com'
     TRIGGER = 'trigger.extracttable.com'
     RESULT = 'getresult.extracttable.com'
+    PRESIGN = 'presign-dev.extracttable.com'
 
 
 class JobStatus:
